requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via -r requirements.in
dependency-injector==4.41.0
    # via -r requirements.in
distlib==0.3.8
    # via virtualenv
filelock==3.14.0
    # via virtualenv
identify==2.5.36
    # via pre-commit
joblib==1.4.0
    # via scikit-learn
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.4
    # via
    #   -r requirements.in
    #   scikit-learn
    #   scipy
platformdirs==4.2.1
    # via virtualenv
pre-commit==3.7.0
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   pre-commit
ruff==0.4.2
    # via -r requirements.in
scikit-learn==1.4.2
    # via -r requirements.in
scipy==1.13.0
    # via scikit-learn
six==1.16.0
    # via dependency-injector
threadpoolctl==3.5.0
    # via scikit-learn
virtualenv==20.26.1
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
